# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Mbarara – Civic Issue Reporting System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  Pending: 'Pending',
  In_Progress: 'In Progress',
  Resolved: 'Resolved',
} as const;

export interface Report {
  id: number;
  /** @nullable */
  name?: string | null;
  phone: string;
  category: string;
  description: string;
  /** @nullable */
  imagePath?: string | null;
  /** @nullable */
  location?: string | null;
  status: ReportStatus;
  department: string;
  createdAt: string;
}

export type ReportInputCategory = typeof ReportInputCategory[keyof typeof ReportInputCategory];


export const ReportInputCategory = {
  Waste_Management: 'Waste Management',
  Health: 'Health',
  Education: 'Education',
  Security: 'Security',
  Water_Supply: 'Water Supply',
  Electricity: 'Electricity',
  'Roads_&_Transport': 'Roads & Transport',
  Other: 'Other',
} as const;

export interface ReportInput {
  name?: string;
  phone: string;
  category: ReportInputCategory;
  description: string;
  imagePath?: string;
  location?: string;
}

export type ReportUpdateStatus = typeof ReportUpdateStatus[keyof typeof ReportUpdateStatus];


export const ReportUpdateStatus = {
  Pending: 'Pending',
  In_Progress: 'In Progress',
  Resolved: 'Resolved',
} as const;

export interface ReportUpdate {
  status?: ReportUpdateStatus;
  description?: string;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  Pending: 'Pending',
  In_Progress: 'In Progress',
  Resolved: 'Resolved',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
}

export interface StatsSummary {
  total: number;
  pending: number;
  inProgress: number;
  resolved: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListReportsParams = {
category?: string;
status?: string;
search?: string;
};