# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Mbarara – Civic Issue Reporting System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reports
    description: Civic issue reports
  - name: stats
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all civic reports
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a new civic report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update a report's status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{id}/status:
    patch:
      operationId: updateReportStatus
      tags: [reports]
      summary: Update only the status of a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Report with updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Aggregated counts and breakdowns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/by-category:
    get:
      operationId: getStatsByCategory
      tags: [stats]
      summary: Get report counts grouped by category
      responses:
        "200":
          description: Counts per category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"

  /stats/by-status:
    get:
      operationId: getStatsByStatus
      tags: [stats]
      summary: Get report counts grouped by status
      responses:
        "200":
          description: Counts per status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"

  /stats/recent:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get the 5 most recently submitted reports
      responses:
        "200":
          description: Recent reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Report:
      type: object
      required: [id, phone, category, description, status, department, createdAt]
      properties:
        id:
          type: integer
        name:
          type: ["string", "null"]
        phone:
          type: string
        category:
          type: string
        description:
          type: string
        imagePath:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [Pending, "In Progress", Resolved]
        department:
          type: string
        createdAt:
          type: string
          format: date-time

    ReportInput:
      type: object
      required: [phone, category, description]
      properties:
        name:
          type: string
        phone:
          type: string
        category:
          type: string
          enum: [Waste Management, Health, Education, Security, Water Supply, Electricity, Roads & Transport, Other]
        description:
          type: string
        imagePath:
          type: string
        location:
          type: string

    ReportUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [Pending, "In Progress", Resolved]
        description:
          type: string

    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Pending, "In Progress", Resolved]

    StatsSummary:
      type: object
      required: [total, pending, inProgress, resolved]
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        resolved:
          type: integer

    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer

    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string