# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Mbarara – Civic Issue Reporting System API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatus } from './reportStatus';

export interface Report {
  id: number;
  /** @nullable */
  name?: string | null;
  phone: string;
  category: string;
  description: string;
  /** @nullable */
  imagePath?: string | null;
  /** @nullable */
  location?: string | null;
  status: ReportStatus;
  department: string;
  createdAt: Date;
}