# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Mbarara – Civic Issue Reporting System API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputCategory } from './reportInputCategory';

export interface ReportInput {
  name?: string;
  phone: string;
  category: ReportInputCategory;
  description: string;
  imagePath?: string;
  location?: string;
}