# lib/api-zod/src/generated/types/statusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Mbarara – Civic Issue Reporting System API
 * OpenAPI spec version: 0.1.0
 */
import type { StatusUpdateStatus } from './statusUpdateStatus';

export interface StatusUpdate {
  status: StatusUpdateStatus;
}